+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#1250** : Fixed logging description for reading and writing documents.
+
 * Issue **#1251** : Copy permissions from a parent now shows changes prior to the user clicking ok.
 
 * Issue **#758** : You no longer need the `Manage Processors` privilege to call `stroom:meta('Pipeline')` in XSLT.